requirements.txt
mmcv
argh==0.26.2
blessings==1.7
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
docker-pycreds==0.4.0
gitdb2==2.0.6
GitPython==3.0.5
gpustat==0.6.0
gql==0.2.0
graphql-core==1.1
idna==2.8
numpy==1.18.1
nltk==3.5
nvidia-ml-py3==7.352.0
pathtools==0.1.2
Pillow==6.1.0
promise==2.3
psutil==5.6.7
python-dateutil==2.8.1
PyYAML==5.3
requests==2.22.0
sentry-sdk==0.14.0
shortuuid==0.5.0
six==1.13.0
smmap2==2.0.5
subprocess32==3.5.4
torch==1.2.0
torchvision==0.4.0
tqdm==4.41.1
urllib3==1.25.7
wandb==0.8.35
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@6b5e8953a80aef5b324104dc0c2e9b8c34d622bd
watchdog==0.9.0
xxhash==1.4.3
